@@ -1,24 +1,21 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Azure.ServiceBus;
-using Microsoft.Azure.ServiceBus.Core;
 using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-using System.Text;
+using Rochambot;
 
 namespace GameMaster
 {
     public class GameMaster : BackgroundService
     {
-        private readonly IConfiguration _configuration;
-        private readonly ILogger<GameMaster> _logger;
-        IMessageSession _scoringSession;
-        QueueClient _sessionResultsClient;
+        readonly IConfiguration _configuration;
+        readonly ILogger<GameMaster> _logger;
+        readonly IMessageSession _scoringSession;
+
+        IQueueClient _sessionResultsClient;
         ISessionClient _playerShapeSelectionClient;
 
         public GameMaster(ILogger<GameMaster> logger, IConfiguration configuration)
@@ -39,36 +36,43 @@ namespace GameMaster
         {
             await _playerShapeSelectionClient.CloseAsync();
             await _sessionResultsClient.CloseAsync();
+
             await base.StopAsync(token);
         }
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
         {
             try
-            {  
+            {
                 var session = await _playerShapeSelectionClient.AcceptMessageSessionAsync();
 
                 while (!stoppingToken.IsCancellationRequested)
                 {
                     var message = await session.ReceiveAsync();
-                    
-                    if(message == null) { continue; }
+                    if (message is null)
+                    {
+                        continue;
+                    }
 
-                    _logger.LogInformation("Message Received {message}", message);
-                    var sessionResult = JsonConvert.DeserializeObject<SessionResult>(
-                        UTF8Encoding.UTF8.GetString(message.Body)
-                    ).Score();
-                    
-                    var respMessage = new Message(UTF8Encoding.UTF8.GetBytes(
-                        JsonConvert.SerializeObject(sessionResult)
-                    ));
+                    _logger.LogInformation($"Message Received {message}");
 
-                    respMessage.SessionId = message.ReplyToSessionId;
-                    await _sessionResultsClient.SendAsync(respMessage);
+                    var sessionResult =
+                        message.Body
+                               .FromUTF8Bytes()
+                               .To<SessionResult>()
+                               .DetermineScore();
+
+                    var responseMessage =
+                        new Message(sessionResult.ToJson().ToUTF8Bytes())
+                        {
+                            SessionId = message.ReplyToSessionId
+                        };
+
+                    await _sessionResultsClient.SendAsync(responseMessage);
                     await session.CompleteAsync(message.SystemProperties.LockToken);
                 }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 _logger.LogError(ex, "GameMaster error");
             }
